third_party/WebKit/LayoutTests/fast/css/invalidation/detach-reattach-shadow-expected.txt

CONSOLE WARNING: line 13: Element.createShadowRoot is deprecated and will be removed in M73, around March 2019. Please use Element.attachShadow instead. See https://www.chromestatus.com/features/4507242028072960 for more details.
PASS successfullyParsed is true

TEST COMPLETE

PASS getComputedStyle(inner).backgroundColor is "rgba(0, 0, 0, 0)"
